airflow/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
faker==19.6.2
pandas==2.1.1